changed some structural properties of the app. Collection view isn't populating.
